+++ programs/dll/uisys_client/src/raw.rs
@@ -1,7 +1,7 @@
 //! Raw export struct definition — must exactly match the DLL's UisysExports layout.
 
 /// Base virtual address where uisys.dll is mapped.
-pub const UISYS_BASE: u32 = 0x0400_0000;
+pub const UISYS_BASE: usize = 0x0400_0000;
 
 #[repr(C)]
 pub struct UisysExports {
