+++ stdlib/src/fs.rs
@@ -9,11 +9,11 @@ pub const O_CREATE: u32 = 4;
 pub const O_TRUNC: u32 = 8;
 
 pub fn write(fd: u32, buf: &[u8]) -> u32 {
-    syscall3(SYS_WRITE, fd, buf.as_ptr() as u32, buf.len() as u32)
+    syscall3(SYS_WRITE, fd as u64, buf.as_ptr() as u64, buf.len() as u64)
 }
 
 pub fn read(fd: u32, buf: &mut [u8]) -> u32 {
-    syscall3(SYS_READ, fd, buf.as_mut_ptr() as u32, buf.len() as u32)
+    syscall3(SYS_READ, fd as u64, buf.as_mut_ptr() as u64, buf.len() as u64)
 }
 
 pub fn open(path: &str, flags: u32) -> u32 {
@@ -21,11 +21,11 @@ pub fn open(path: &str, flags: u32) -> u32 {
     let len = path.len().min(256);
     buf[..len].copy_from_slice(&path.as_bytes()[..len]);
     buf[len] = 0;
-    syscall3(SYS_OPEN, buf.as_ptr() as u32, flags, 0)
+    syscall3(SYS_OPEN, buf.as_ptr() as u64, flags as u64, 0)
 }
 
 pub fn close(fd: u32) -> u32 {
-    syscall1(SYS_CLOSE, fd)
+    syscall1(SYS_CLOSE, fd as u64)
 }
 
 /// Read directory entries. Returns number of entries (or u32::MAX on error).
@@ -35,7 +35,7 @@ pub fn readdir(path: &str, buf: &mut [u8]) -> u32 {
     let len = path.len().min(256);
     path_buf[..len].copy_from_slice(&path.as_bytes()[..len]);
     path_buf[len] = 0;
-    syscall3(SYS_READDIR, path_buf.as_ptr() as u32, buf.as_mut_ptr() as u32, buf.len() as u32)
+    syscall3(SYS_READDIR, path_buf.as_ptr() as u64, buf.as_mut_ptr() as u64, buf.len() as u64)
 }
 
 /// Get file status. Returns 0 on success. Writes [type:u32, size:u32] to buf.
@@ -44,7 +44,7 @@ pub fn stat(path: &str, stat_buf: &mut [u32; 2]) -> u32 {
     let len = path.len().min(256);
     path_buf[..len].copy_from_slice(&path.as_bytes()[..len]);
     path_buf[len] = 0;
-    syscall2(SYS_STAT, path_buf.as_ptr() as u32, stat_buf.as_mut_ptr() as u32)
+    syscall2(SYS_STAT, path_buf.as_ptr() as u64, stat_buf.as_mut_ptr() as u64)
 }
 
 /// Create a directory. Returns 0 on success, u32::MAX on error.
@@ -53,7 +53,7 @@ pub fn mkdir(path: &str) -> u32 {
     let len = path.len().min(256);
     buf[..len].copy_from_slice(&path.as_bytes()[..len]);
     buf[len] = 0;
-    syscall1(SYS_MKDIR, buf.as_ptr() as u32)
+    syscall1(SYS_MKDIR, buf.as_ptr() as u64)
 }
 
 /// Delete a file. Returns 0 on success, u32::MAX on error.
@@ -62,7 +62,7 @@ pub fn unlink(path: &str) -> u32 {
     let len = path.len().min(256);
     buf[..len].copy_from_slice(&path.as_bytes()[..len]);
     buf[len] = 0;
-    syscall1(SYS_UNLINK, buf.as_ptr() as u32)
+    syscall1(SYS_UNLINK, buf.as_ptr() as u64)
 }
 
 /// Truncate a file to zero length. Returns 0 on success, u32::MAX on error.
@@ -71,7 +71,7 @@ pub fn truncate(path: &str) -> u32 {
     let len = path.len().min(256);
     buf[..len].copy_from_slice(&path.as_bytes()[..len]);
     buf[len] = 0;
-    syscall1(SYS_TRUNCATE, buf.as_ptr() as u32)
+    syscall1(SYS_TRUNCATE, buf.as_ptr() as u64)
 }
 
 // Seek whence constants
@@ -81,21 +81,21 @@ pub const SEEK_END: u32 = 2;
 
 /// Seek within an open file. Returns new position or u32::MAX on error.
 pub fn lseek(fd: u32, offset: i32, whence: u32) -> u32 {
-    syscall3(SYS_LSEEK, fd, offset as u32, whence)
+    syscall3(SYS_LSEEK, fd as u64, offset as i64 as u64, whence as u64)
 }
 
 /// Get file information by fd. Returns 0 on success.
 /// Writes [type:u32, size:u32, position:u32] to stat_buf.
 pub fn fstat(fd: u32, stat_buf: &mut [u32; 3]) -> u32 {
-    syscall2(SYS_FSTAT, fd, stat_buf.as_mut_ptr() as u32)
+    syscall2(SYS_FSTAT, fd as u64, stat_buf.as_mut_ptr() as u64)
 }
 
 /// Get current working directory. Returns length or u32::MAX on error.
 pub fn getcwd(buf: &mut [u8]) -> u32 {
-    syscall2(SYS_GETCWD, buf.as_mut_ptr() as u32, buf.len() as u32)
+    syscall2(SYS_GETCWD, buf.as_mut_ptr() as u64, buf.len() as u64)
 }
 
 /// Check if fd refers to a terminal. Returns 1 for tty, 0 otherwise.
 pub fn isatty(fd: u32) -> u32 {
-    syscall1(SYS_ISATTY, fd)
+    syscall1(SYS_ISATTY, fd as u64)
 }
